+++ crates/sui-indexer/src/main.rs
@@ -2,9 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use clap::Parser;
+use tracing::info;
+
 use sui_indexer::errors::IndexerError;
 use sui_indexer::store::PgIndexerStore;
-use sui_indexer::{new_pg_connection_pool, Indexer, IndexerConfig};
+use sui_indexer::utils::reset_database;
+use sui_indexer::{get_pg_pool_connection, new_pg_connection_pool, Indexer, IndexerConfig};
 use sui_node::metrics::start_prometheus_server;
 
 #[tokio::main]
@@ -15,6 +18,7 @@ async fn main() -> Result<(), IndexerError> {
         .init();
 
     let indexer_config = IndexerConfig::parse();
+    info!("indexer config: {:#?}", indexer_config);
     let registry_service = start_prometheus_server(
         // NOTE: this parses the input host addr and port number for socket addr,
         // so unwrap() is safe here.
@@ -28,6 +32,16 @@ async fn main() -> Result<(), IndexerError> {
 
     let registry = registry_service.default_registry();
     let pg_connection_pool = new_pg_connection_pool(&indexer_config.db_url).await?;
+    if indexer_config.reset_db {
+        let mut conn = get_pg_pool_connection(&pg_connection_pool)?;
+        reset_database(&mut conn, /* drop_all */ true).map_err(|e| {
+            IndexerError::PostgresResetError(format!(
+                "unable to reset database with url: {:?} and err: {:?}",
+                indexer_config.db_url.clone(),
+                e
+            ))
+        })?;
+    }
     let store = PgIndexerStore::new(pg_connection_pool);
 
     Indexer::start(&indexer_config, &registry, store).await
